// code/frontend/src/app/services/models/page-real-estate-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PageableObject } from '../models/pageable-object';
import { RealEstateResponse } from '../models/real-estate-response';
import { SortObject } from '../models/sort-object';
export interface PageRealEstateResponse {
  content?: Array<RealEstateResponse>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  pageable?: PageableObject;
  size?: number;
  sort?: SortObject;
  totalElements?: number;
  totalPages?: number;
}